Add MSRV build to CI
We were not testing MSRV 1.56.1 before due to a build failure. This
allowed a bit of feature creap which relies on newer versions of
Rust. Luckily, it was a single feature in our unit tests, so easy
to refactor. Now enforcing the MSRV in CI, but this did require us
to use precise versions of memchr and backtrace. Currently controlled
by our lock file.