feat: add decrypt_in_place to outbound cipher
The BIP-324 protocol is packet in nature, not stream based, and the
encrypted packets are not the same size as plaintext. They get length
bits added to the front and authentication bits tacked on the back. So
encryption in place would be a huge pain. And generally decryption in
place is weird too, because there are these bytes on either end which
no longer make sense to the plaintext. However, there is a solid use
case for *decryption* in place in the handshake where packet content
is not really cared about.