protocol: Remove explicit `rand` dependency
We only bring in `rand` explicitly for the `Rng` trait, coming from
`RngCore`. The `Rng` trait is a nightmare to implement, and we only use
a single call to `.fill`. If we export a far simpler trait that
represents filling a 32 byte slice, then users can have a far easier
time implementing this trait. This has come up in the past with this
project: https://github.com/issues/recent?issue=alfred-hodler%7Cpushtx%7C3

This also removes the need for `rand` as a dependency at all, and it may
be brought in through `bitcoin` (or later `secp256k1`)