traffic: Remove the use of `bip324::serde`
To prepare for removing `bitcoin` as a dependency, we can get a jump by
removing the use of `bip324::serde` in the test suite. Understandable if
it is too weird to depend on specific commits in the test suite, but
throwing it up because it is not a horribly complicated change.

Co-authored-by: Nick Johnson <nick@yonson.dev>