protocol: Remove `bitcoin` for `secp256k1`
To help with type safety a bit, I have public callsites use an `impl
Borrow<[u8; 4]>`, which the `p2p::Magic` type implements. This way one
can pass magic variables directly. In the handshake this just gets
dereferenced anyway, but it still feels more ergonomic than forcing a
`to_bytes` on the caller IMO.

I keep the `secp256k1` the same, but maybe it can be bumped. I forget if
there are problems when attempting to build conflicting versions or not.

Concat usage was dropped in favor of explicit array handling. This was
probably being iron'd out by the compiler before, but this patch
introduces Borrow which broke no-std environment support.

Co-authored-by: Nick Johnson <nick@yonson.dev>